+++ js/main.js
@@ -1,3 +1,3 @@
-import { getMockedPhotos, PHOTOS_LENGTH } from './photos.js';
+import { getMockedPhotos, PHOTOS_LENGTH } from './mock-data.js';
 
 getMockedPhotos(PHOTOS_LENGTH);
